Renamed Files and Cleaned Code
