+++ rr_iarrc/CMakeLists.txt
@@ -66,4 +66,4 @@ add_subdirectory(src/cone_detection)
 add_subdirectory(src/urc_controller)
 add_subdirectory(src/turning_test)
 add_subdirectory(src/sign_detector)
-add_subdirectory(src/urc_lane_follower)
+add_subdirectory(src/urc_image_merger)
