+++ rr_iarrc/src/cone_detection/CMakeLists.txt
@@ -6,4 +6,9 @@ target_link_libraries(cone_bottom_detector ${catkin_LIBRARIES})
 target_link_libraries( cone_bottom_detector ${OpenCV_LIBS} )
 add_dependencies(cone_bottom_detector rr_platform_generate_messages_cpp)
 
+add_executable(cone_height_detector cone_height_detector.cpp)
+target_link_libraries(cone_height_detector ${catkin_LIBRARIES})
+target_link_libraries( cone_height_detector ${OpenCV_LIBS} )
+add_dependencies(cone_height_detector rr_platform_generate_messages_cpp)
+
