+++ rr_iarrc/src/cone_detection/cone_height_detector.cpp
