+++ rr_iarrc/src/drag_centerline_planner/CMakeLists.txt
@@ -8,7 +8,7 @@ target_link_libraries(drag_centerline_planner PID ${catkin_LIBRARIES})
 target_link_libraries( drag_centerline_planner ${OpenCV_LIBS} )
 add_dependencies(drag_centerline_planner rr_platform_generate_messages_cpp)
 
-add_executable(drag_centerline_planner_daniel drag_centerline_planner_daniel.cpp)
-target_link_libraries(drag_centerline_planner_daniel PID ${catkin_LIBRARIES})
-target_link_libraries( drag_centerline_planner_daniel ${OpenCV_LIBS} )
-add_dependencies(drag_centerline_planner_daniel rr_platform_generate_messages_cpp)
+add_executable(drag_centerline_planner_contours drag_centerline_planner_contours.cpp)
+target_link_libraries(drag_centerline_planner_contours PID ${catkin_LIBRARIES})
+target_link_libraries( drag_centerline_planner_contours ${OpenCV_LIBS} )
+add_dependencies(drag_centerline_planner_contours rr_platform_generate_messages_cpp)
